backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
firebase-admin==6.0.1
google-cloud-texttospeech==2.12.0
google-cloud-speech==2.20.0
google-cloud-language==2.11.0
twilio==8.0.0
python-dotenv==1.0.0
APScheduler==3.10.1
gunicorn==20.1.0
requests==2.31.0
google-cloud-translate
google-auth
python-dateutil
apscheduler==3.10.4
pytz==2023.3
pydub
supabase